refactor(scrappedEvent): ScrappedEvent를 Event로 통합
