+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/service/EventService.kt
@@ -1,8 +1,12 @@
 package com.emmsale.data.service
 
+import com.emmsale.data.apiModel.request.ScrappedEventCreateRequest
 import com.emmsale.data.apiModel.response.EventResponse
 import com.emmsale.data.common.retrofit.callAdapter.ApiResponse
+import retrofit2.http.Body
+import retrofit2.http.DELETE
 import retrofit2.http.GET
+import retrofit2.http.POST
 import retrofit2.http.Path
 import retrofit2.http.Query
 
@@ -39,4 +43,17 @@ interface EventService {
         @Query("statuses") statuses: List<String> = emptyList(),
         @Query("category") category: String? = null,
     ): ApiResponse<List<EventResponse>>
+
+    @GET("/scraps")
+    suspend fun getScrappedEvents(): ApiResponse<List<EventResponse>>
+
+    @POST("/scraps")
+    suspend fun scrapEvent(
+        @Body scrappedEventCreateRequest: ScrappedEventCreateRequest,
+    ): ApiResponse<Unit>
+
+    @DELETE("/scraps")
+    suspend fun deleteScrap(
+        @Query("event-id") eventId: Long,
+    ): ApiResponse<Unit>
 }
