+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/scrappedEventList/ScrappedEventFragment.kt
@@ -47,6 +47,6 @@ class ScrappedEventFragment : BaseFragment<FragmentScrappedEventBinding>() {
     }
 
     private fun showEventDetail(scrappedEventUiState: ScrappedEventUiState) {
-        EventDetailActivity.startActivity(requireContext(), scrappedEventUiState.scrappedEvent.id)
+        EventDetailActivity.startActivity(requireContext(), scrappedEventUiState.event.id)
     }
 }
