+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/scrappedEventList/ScrappedEventViewModel.kt
@@ -3,7 +3,7 @@ package com.emmsale.presentation.ui.scrappedEventList
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.emmsale.data.common.retrofit.callAdapter.Success
-import com.emmsale.data.repository.interfaces.ScrappedEventRepository
+import com.emmsale.data.repository.interfaces.EventRepository
 import com.emmsale.presentation.common.FetchResult
 import com.emmsale.presentation.common.livedata.NotNullLiveData
 import com.emmsale.presentation.common.livedata.NotNullMutableLiveData
@@ -15,7 +15,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 class ScrappedEventViewModel @Inject constructor(
-    private val scrappedEventRepository: ScrappedEventRepository,
+    private val eventRepository: EventRepository,
 ) : ViewModel(), Refreshable {
     private val _scrappedEvents = NotNullMutableLiveData(ScrappedEventsUiState())
     val scrappedEvents: NotNullLiveData<ScrappedEventsUiState> = _scrappedEvents
@@ -23,7 +23,7 @@ class ScrappedEventViewModel @Inject constructor(
     override fun refresh() {
         changeToLoadingState()
         viewModelScope.launch {
-            when (val response = scrappedEventRepository.getScrappedEvents()) {
+            when (val response = eventRepository.getScrappedEvents()) {
                 is Success -> _scrappedEvents.value = ScrappedEventsUiState.from(response.data)
                 else -> changeToErrorState()
             }
