+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/scrappedEventList/uiState/ScrappedEventsUiState.kt
@@ -1,6 +1,6 @@
 package com.emmsale.presentation.ui.scrappedEventList.uiState
 
-import com.emmsale.data.model.ScrappedEvent
+import com.emmsale.data.model.Event
 import com.emmsale.presentation.common.FetchResult
 import com.emmsale.presentation.common.FetchResultUiState
 
@@ -9,7 +9,7 @@ data class ScrappedEventsUiState(
     override val fetchResult: FetchResult = FetchResult.LOADING,
 ) : FetchResultUiState() {
     companion object {
-        fun from(scrappedEvents: List<ScrappedEvent>): ScrappedEventsUiState =
+        fun from(scrappedEvents: List<Event>): ScrappedEventsUiState =
             ScrappedEventsUiState(
                 list = scrappedEvents.map { ScrappedEventUiState.from(it) },
                 fetchResult = FetchResult.SUCCESS,
