+++ android/2023-emmsale/app/src/main/res/layout/item_scrapped_event.xml
@@ -37,7 +37,7 @@
             android:layout_width="0dp"
             android:layout_height="0dp"
             android:contentDescription="@string/event_poster_description"
-            app:imageUrl="@{event.scrappedEvent.posterUrl}"
+            app:imageUrl="@{event.event.posterImageUrl}"
             app:layout_constraintDimensionRatio="326:180"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
@@ -52,11 +52,11 @@
             android:layout_marginTop="12dp"
             android:textSize="13sp"
             android:textStyle="bold"
-            app:eventApplyingStatus="@{event.scrappedEvent.applicationStatus}"
+            app:eventApplyingStatus="@{event.event.applicationStatus}"
             app:layout_constraintStart_toStartOf="@+id/iv_event_poster"
             app:layout_constraintTop_toBottomOf="@+id/iv_event_poster"
             app:layout_goneMarginTop="9dp"
-            app:visible="@{!(event.scrappedEvent.progressStatus instanceof EventProgressStatus.Ended)}"
+            app:visible="@{!(event.event.progressStatus instanceof EventProgressStatus.Ended)}"
             tools:text="D-21"
             tools:textColor="@color/primary_color" />
 
@@ -65,7 +65,7 @@
             android:layout_width="1dp"
             android:layout_height="0dp"
             android:layout_marginStart="6dp"
-            app:visible="@{!(event.scrappedEvent.progressStatus instanceof EventProgressStatus.Ended)}"
+            app:visible="@{!(event.event.progressStatus instanceof EventProgressStatus.Ended)}"
             android:background="@color/light_gray"
             app:layout_constraintBottom_toBottomOf="@+id/tv_application_status"
             app:layout_constraintStart_toEndOf="@+id/tv_application_status"
@@ -80,7 +80,7 @@
             android:textSize="13sp"
             android:textStyle="bold"
             app:layout_goneMarginStart="0dp"
-            app:eventProgressStatus="@{event.scrappedEvent.progressStatus}"
+            app:eventProgressStatus="@{event.event.progressStatus}"
             app:layout_constraintStart_toEndOf="@+id/divider_event_status"
             app:layout_constraintTop_toBottomOf="@+id/iv_event_poster"
             tools:text="D-21"
@@ -90,7 +90,7 @@
             android:id="@+id/tv_event_name"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:text="@{event.scrappedEvent.name}"
+            android:text="@{event.event.name}"
             android:textColor="@color/black"
             android:layout_marginTop="9dp"
             android:textSize="15sp"
@@ -108,7 +108,7 @@
             android:textSize="12sp"
             app:layout_constraintStart_toStartOf="@+id/iv_event_poster"
             app:layout_constraintTop_toBottomOf="@+id/tv_event_name"
-            app:paymentType="@{event.scrappedEvent.paymentType}"
+            app:paymentType="@{event.event.paymentType}"
             tools:text="무료" />
 
         <View
@@ -132,7 +132,7 @@
             app:layout_constraintStart_toStartOf="@+id/divider"
             app:layout_constraintTop_toBottomOf="@+id/tv_event_name"
             app:layout_constraintTop_toTopOf="@+id/tv_pay_type"
-            app:onOfflineMode="@{event.scrappedEvent.onOfflineMode}"
+            app:onOfflineMode="@{event.event.onOfflineMode}"
             tools:text="온라인" />
 
         <com.google.android.material.chip.ChipGroup
@@ -141,7 +141,7 @@
             android:layout_height="wrap_content"
             android:layout_marginTop="24dp"
             android:layout_marginBottom="10dp"
-            app:eventChips="@{event.scrappedEvent.tags}"
+            app:eventChips="@{event.event.tags}"
             app:layout_constraintEnd_toEndOf="@+id/iv_event_poster"
             app:layout_constraintStart_toStartOf="@+id/iv_event_poster"
             app:layout_constraintTop_toBottomOf="@+id/tv_pay_type" />
